backend/requirements.txt
zhipuai==1.0.7
requests==2.31.0
Pillow==10.0.0
python-dotenv==1.0.0
fastapi==0.109.0
uvicorn==0.25.0
python-multipart==0.0.6